refactor: split project build helpers
